apply routers, controllers division logic
