+++ my-postman/routes/viewRoutes.js
@@ -1,0 +1,12 @@
+const express = require('express');
+
+const {
+	getOverview,
+} = require('../controllers/viewsController');
+
+const router = express.Router();
+
+router.route('/').get(getOverview);
+	
+
+module.exports = router;
