+++ my-postman/utils/cors.js
@@ -1,0 +1,11 @@
+const customCors = (req, res, next) => {
+	res.header('Access-Control-Allow-Origin', '*');
+	res.header(
+		'Access-Control-Allow-Headers',
+		'Origin, X-Requested-With, Content-Type, Accept',
+	);
+	res.header('Access-Control-Allow-Methods', 'GET, POST, PUT, DELETE, OPTIONS');
+	next();
+};
+
+module.exports = customCors;
